Introduced proper use of validation data.
Training data is now back to being state-class pairs. Validation and test data is state-distribution pairs, where the distribution is the empirical distribution over classes associated with that state in the data set. I don't 100% know if this is a valid move to make. It messes up the validation and test scores a heck of a lot. Everything is now cross entropy loss instead of accuracy, since we're comparing distributions.